requirements.txt
Django==4.2.7
Pillow>=10.2.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
lxml==5.1.0
beautifulsoup4==4.12.2
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
django-bootstrap4==23.2
django-allauth==0.57.0 
requests>=2.31.0